+++ src/core/study_processor.ts
